Fix CI flows requiring vscode latest
The update applies only to the CI flows requiring to build the latest
version of vscode which requires node version > 14.

The two remaining CI flows are left using node version 12 as they are normally
used to rebuild with older versions of vscode for the time being.

it additionally:
* Upgrades yarn dependecies which brings latest theia as the
example application.
* Cleans up dependencies by removing @types/htmlparser2,
and unifies the version for @types/react.
* Updates the vscode submodule to version 1.58.2 which requires node v14.

Co-Authored-By: Paul Maréchal <paul.marechal@ericsson.com>
Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>